requirements.txt
pymodbus==3.6.2
flask==2.3.3
flask-socketio==5.3.6
influxdb-client==1.38.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.4
Pillow==10.1.0